v1.4 Introduce FileChooserEventArgs.Page and .SetFilesAsync
